DeepConcolic/requirements-deepconcolic.txt
numpy==1.19.5
scipy==1.4.1
tensorflow\>=2.4
pomegranate==0.14
scikit-learn 
scikit-image 
pulp 
keract 
np_utils 
adversarial-robustness-toolbox 
parse 
tabulate 
pysmt 
saxpy 
keras 
menpo 
patool 
z3-solver 
pyvis
pymongo
python-dotenv